be/fastapi-rag/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pypdf
ollama